requirements.txt
HTMLTestRunner
PyMySQL
Appium-Python-Client
selenium
PyYAML